+++ service-worker.js
@@ -2,15 +2,17 @@ self.addEventListener('install', event => {
   event.waitUntil(
     caches.open('pick-me-a-game-v1').then(cache => {
       return cache.addAll([
-        '/',
-        '/index.html',
-        '/src/js/index.js',
-        '/src/css/index.css',
-        '/src/css/achievements.css',
-        '/src/css/header.css',
-        '/src/css/footer.css',
-        '/src/js/addyourgame_toggle.js',
-        '/src/js/achievements.js',
+        './',
+        './index.html',
+        './src/js/index.js',
+        './src/css/index.css',
+        './src/css/achievements.css',
+        './src/css/header.css',
+        './src/css/footer.css',
+        './src/js/addyourgame_toggle.js',
+        './src/js/achievements.js',
+        './src/img/background-elden.jpg',
+        './src/img/background_footer.jpg',
         // add other assets as needed
       ]);
     })